Add retries to default file tree view
